Moved submission tests to new test file
