+++ database/gormdb_submission_test.go
@@ -1,0 +1,319 @@
+package database_test
+
+import (
+	"fmt"
+	"reflect"
+	"testing"
+
+	pb "github.com/autograde/quickfeed/ag"
+	"github.com/jinzhu/gorm"
+)
+
+func TestGormDBGetSubmissionForUser(t *testing.T) {
+	db, cleanup := setup(t)
+	defer cleanup()
+	query := &pb.Submission{AssignmentID: 10, UserID: 10}
+	if _, err := db.GetSubmission(query); err != gorm.ErrRecordNotFound {
+		t.Errorf("have error '%v' wanted '%v'", err, gorm.ErrRecordNotFound)
+	}
+}
+
+func TestGormDBUpdateSubmission(t *testing.T) {
+	db, cleanup := setup(t)
+	defer cleanup()
+
+	// when we create a new submission for the same course lab and user, it will update the old one,
+	// instead of creating an extra record
+	// check that it is still approved after using create method
+
+	teacher := createFakeUser(t, db, 10)
+	// create a course and an assignment
+	var course pb.Course
+	if err := db.CreateCourse(teacher.ID, &course); err != nil {
+		t.Fatal(err)
+	}
+	assignment := pb.Assignment{
+		CourseID: course.ID,
+		Order:    1,
+	}
+	if err := db.CreateAssignment(&assignment); err != nil {
+		t.Fatal(err)
+	}
+
+	// create user and enroll as student
+	user := createFakeUser(t, db, 11)
+	if err := db.CreateEnrollment(&pb.Enrollment{
+		UserID:   user.ID,
+		CourseID: course.ID,
+	}); err != nil {
+		t.Fatal(err)
+	}
+	query := &pb.Enrollment{
+		UserID:   user.ID,
+		CourseID: course.ID,
+		Status:   pb.Enrollment_STUDENT,
+	}
+	if err := db.UpdateEnrollment(query); err != nil {
+		t.Fatal(err)
+	}
+
+	// create another submission for the assignment; now it should succeed
+	if err := db.CreateSubmission(&pb.Submission{
+		AssignmentID: assignment.ID,
+		UserID:       user.ID,
+	}); err != nil {
+		t.Fatal(err)
+	}
+
+	// confirm that the submission is in the database
+	submissions, err := db.GetLastSubmissions(course.ID, &pb.Submission{UserID: user.ID})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(submissions) != 1 {
+		t.Fatalf("have %d submissions want %d", len(submissions), 1)
+	}
+	want := &pb.Submission{
+		ID:           submissions[0].ID,
+		AssignmentID: assignment.ID,
+		UserID:       user.ID,
+		Status:       pb.Submission_NONE,
+		Reviews:      []*pb.Review{},
+	}
+	if !reflect.DeepEqual(submissions[0], want) {
+		t.Errorf("have %#v want %#v", submissions[0], want)
+	}
+
+	if submissions[0].GetStatus() != pb.Submission_NONE {
+		t.Errorf("expected submission to be 'not-approved' but got 'approved'")
+	}
+
+	// approved must stay false
+	err = db.UpdateSubmission(submissions[0])
+	if err != nil {
+		t.Fatal(err)
+	}
+	submissions, err = db.GetLastSubmissions(course.ID, &pb.Submission{UserID: user.ID})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if submissions[0].GetStatus() != pb.Submission_NONE {
+		t.Errorf("expected submission to be 'not-approved' but got 'approved'")
+	}
+	submissions[0].Status = pb.Submission_APPROVED
+	err = db.UpdateSubmission(submissions[0])
+	if err != nil {
+		t.Fatal(err)
+	}
+	submissions, err = db.GetLastSubmissions(course.ID, &pb.Submission{UserID: user.ID})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if submissions[0].GetStatus() != pb.Submission_APPROVED {
+		t.Errorf("expected submission to be 'approved' but got 'not-approved'")
+	}
+}
+
+func TestGormDBGetNonExistingSubmissions(t *testing.T) {
+	db, cleanup := setup(t)
+	defer cleanup()
+	if _, err := db.GetLastSubmissions(10, &pb.Submission{UserID: 10}); err != gorm.ErrRecordNotFound {
+		t.Errorf("have error '%v' wanted '%v'", err, gorm.ErrRecordNotFound)
+	}
+}
+
+func TestGormDBInsertSubmissions(t *testing.T) {
+	db, cleanup := setup(t)
+	defer cleanup()
+
+	if err := db.CreateSubmission(&pb.Submission{
+		AssignmentID: 1,
+		UserID:       1,
+	}); err != gorm.ErrRecordNotFound {
+		t.Fatal(err)
+	}
+
+	teacher := createFakeUser(t, db, 10)
+	// create a course and an assignment
+	var course pb.Course
+	if err := db.CreateCourse(teacher.ID, &course); err != nil {
+		t.Fatal(err)
+	}
+	assignment := pb.Assignment{
+		CourseID: course.ID,
+		Order:    1,
+	}
+	if err := db.CreateAssignment(&assignment); err != nil {
+		t.Fatal(err)
+	}
+
+	// create a submission for the assignment; should fail
+	if err := db.CreateSubmission(&pb.Submission{
+		AssignmentID: assignment.ID,
+		UserID:       2,
+	}); err != gorm.ErrRecordNotFound {
+		t.Fatal(err)
+	}
+
+	// create user and enroll as student
+	user := createFakeUser(t, db, 11)
+	if err := db.CreateEnrollment(&pb.Enrollment{
+		UserID:   user.ID,
+		CourseID: course.ID,
+	}); err != nil {
+		t.Fatal(err)
+	}
+	query := &pb.Enrollment{
+		UserID:   user.ID,
+		CourseID: course.ID,
+		Status:   pb.Enrollment_STUDENT,
+	}
+	if err := db.UpdateEnrollment(query); err != nil {
+		t.Fatal(err)
+	}
+
+	// create another submission for the assignment; now it should succeed
+	if err := db.CreateSubmission(&pb.Submission{
+		AssignmentID: assignment.ID,
+		UserID:       user.ID,
+	}); err != nil {
+		t.Fatal(err)
+	}
+
+	// confirm that the submission is in the database
+	submissions, err := db.GetLastSubmissions(course.ID, &pb.Submission{UserID: user.ID})
+	if err != nil {
+		t.Fatal(err)
+	}
+	if len(submissions) != 1 {
+		t.Fatalf("have %d submissions want %d", len(submissions), 1)
+	}
+	want := &pb.Submission{
+		ID:           submissions[0].ID,
+		AssignmentID: assignment.ID,
+		UserID:       user.ID,
+		Reviews:      []*pb.Review{},
+	}
+	if !reflect.DeepEqual(submissions[0], want) {
+		t.Errorf("have %#v want %#v", submissions[0], want)
+	}
+}
+
+func TestGormDBGetInsertSubmissions(t *testing.T) {
+	db, cleanup := setup(t)
+	defer cleanup()
+
+	teacher := createFakeUser(t, db, 10)
+	// Create course c1 and c2
+	c1 := pb.Course{OrganizationID: 1}
+	if err := db.CreateCourse(teacher.ID, &c1); err != nil {
+		t.Fatal(err)
+	}
+	c2 := pb.Course{OrganizationID: 2}
+	if err := db.CreateCourse(teacher.ID, &c2); err != nil {
+		t.Fatal(err)
+	}
+
+	// create user and enroll as student
+	user := createFakeUser(t, db, 11)
+
+	// enroll student in course c1
+	if err := db.CreateEnrollment(&pb.Enrollment{
+		UserID:   user.ID,
+		CourseID: c1.ID,
+	}); err != nil {
+		t.Fatal(err)
+	}
+	query := &pb.Enrollment{
+		UserID:   user.ID,
+		CourseID: c1.ID,
+		Status:   pb.Enrollment_STUDENT,
+	}
+	if err := db.UpdateEnrollment(query); err != nil {
+		t.Fatal(err)
+	}
+
+	// Create some assignments
+	assignment1 := pb.Assignment{
+		Order:    1,
+		CourseID: c1.ID,
+	}
+	if err := db.CreateAssignment(&assignment1); err != nil {
+		t.Fatal(err)
+	}
+	assignment2 := pb.Assignment{
+		Order:    2,
+		CourseID: c1.ID,
+	}
+	if err := db.CreateAssignment(&assignment2); err != nil {
+		t.Fatal(err)
+	}
+	assignment3 := pb.Assignment{
+		Order:    1,
+		CourseID: c2.ID,
+	}
+	if err := db.CreateAssignment(&assignment3); err != nil {
+		t.Fatal(err)
+	}
+
+	// Create some submissions. We need IDs set here to be able
+	// to compare local submission structs with database structs.
+	submission1 := pb.Submission{
+		UserID:       user.ID,
+		AssignmentID: assignment1.ID,
+		Reviews:      []*pb.Review{},
+	}
+	if err := db.CreateSubmission(&submission1); err != nil {
+		t.Fatal(err)
+	}
+	submission2 := pb.Submission{
+		ID:           1,
+		UserID:       user.ID,
+		AssignmentID: assignment1.ID,
+		Reviews:      []*pb.Review{},
+	}
+	if err := db.CreateSubmission(&submission2); err != nil {
+		t.Fatal(err)
+	}
+	submission3 := pb.Submission{
+		ID:           2,
+		UserID:       user.ID,
+		AssignmentID: assignment2.ID,
+		Reviews:      []*pb.Review{},
+	}
+	if err := db.CreateSubmission(&submission3); err != nil {
+		t.Fatal(err)
+	}
+
+	// Even if there is three submission, only the latest for each assignment should be returned
+
+	submissions, err := db.GetLastSubmissions(c1.ID, &pb.Submission{UserID: user.ID})
+	if err != nil {
+		t.Fatal(err)
+	}
+	want := []*pb.Submission{&submission2, &submission3}
+	if !reflect.DeepEqual(submissions, want) {
+		fmt.Println("Submissions in the database:")
+		for _, s := range submissions {
+			fmt.Printf("%+v\n", s)
+		}
+		fmt.Println("Expected submissions:")
+		for _, s := range want {
+			fmt.Printf("%+v\n", s)
+		}
+		t.Errorf("have %#v want %#v", submissions, want)
+	}
+	data, err := db.GetLastSubmissions(c1.ID, &pb.Submission{UserID: user.ID})
+	if err != nil {
+		t.Fatal(err)
+	} else if len(data) != 2 {
+		t.Errorf("Expected '%v' elements in the array, got '%v'", 2, len(data))
+	}
+	// Since there is no submissions, but the course and user exist, an empty array should be returned
+	data, err = db.GetLastSubmissions(c2.ID, &pb.Submission{UserID: user.ID})
+	if err != nil {
+		t.Fatal(err)
+	} else if len(data) != 0 {
+		t.Errorf("Expected '%v' elements in the array, got '%v'", 0, len(data))
+	}
+}
